# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "API Minimarket",
                 "description":  "DocumentaciÃ³n de la API del sistema Minimarket",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
                  "/api/usuarios/{id}":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Usuario"
                                                                  ],
                                                         "summary":  "Obtener usuario por ID",
                                                         "description":  "Devuelve los detalles de un usuario especÃ­fico",
                                                         "operationId":  "obtenerUsuarioPorId",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Usuario encontrado",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/EntityModelUsuarioResponseDTO"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Usuario no encontrado",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/EntityModelUsuarioResponseDTO"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "put":  {
                                                         "tags":  [
                                                                      "Usuario"
                                                                  ],
                                                         "summary":  "Actualizar usuario",
                                                         "description":  "Actualiza los datos de un usuario existente",
                                                         "operationId":  "actualizarUsuario",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/Usuario"
                                                                                                                                 }
                                                                                                                  }
                                                                                         },
                                                                             "required":  true
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Usuario actualizado",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/EntityModelUsuarioResponseDTO"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Usuario no encontrado",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/EntityModelUsuarioResponseDTO"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "delete":  {
                                                            "tags":  [
                                                                         "Usuario"
                                                                     ],
                                                            "summary":  "Eliminar usuario",
                                                            "description":  "Elimina un usuario por su ID",
                                                            "operationId":  "eliminarUsuario",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int64"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Usuario eliminado",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/EntityModelMapStringString"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      },
                                                                              "404":  {
                                                                                          "description":  "Usuario no encontrado",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/EntityModelMapStringString"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                         },
...
                  "/api/ventas/checkout/usuario/{usuarioId}":  {
                                                                   "post":  {
                                                                                "tags":  [
                                                                                             "venta-controller"
                                                                                         ],
                                                                                "operationId":  "procesarVentaDesdeCarrito",
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "usuarioId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "integer",
                                                                                                                      "format":  "int64"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK",
                                                                                                              "content":  {
                                                                                                                              "*/*":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/Venta"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                                          }
                                                                                                          }
                                                                                              }
                                                                            }
                                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "EntityModelUsuarioResponseDTO":  {
                                                                             "type":  "object",
                                                                             "properties":  {
                                                                                                "id":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int64"
                                                                                                       },
                                                                                                "username":  {
                                                                                                                 "type":  "string"
                                                                                                             },
                                                                                                "roles":  {
                                                                                                              "type":  "array",
                                                                                                              "items":  {
                                                                                                                            "type":  "string"
                                                                                                                        },
                                                                                                              "uniqueItems":  true
                                                                                                          },
                                                                                                "_links":  {
                                                                                                               "$ref":  "#/components/schemas/Links"
                                                                                                           }
                                                                                            }
                                                                         },
...
                                       "EntityModelMapStringString":  {
                                                                          "type":  "object",
                                                                          "properties":  {
                                                                                             "_links":  {
                                                                                                            "$ref":  "#/components/schemas/Links"
                                                                                                        }
                                                                                         }
                                                                      },
                                       "Links":  {
                                                     "type":  "object",
                                                     "additionalProperties":  {
                                                                                  "$ref":  "#/components/schemas/Link"
                                                                              }
                                                 }
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "AutenticaciÃ³n JWT usando Bearer Token",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}